model/requirements.txt
pandas
numpy
matplotlib
seaborn
pytz
pytest
mock
requests
sklearn
fluent-logger
msgpack
fastparquet
schedule